backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
assemblyai
requests
python-dotenv
email-validator
